refractor and error handling

+++ controllers/Delete.js
@@ -6,6 +6,10 @@ const path = require('path');
 const Delete = async (id) => {
     const filePath = path.join(__dirname + '/authStorage/authToken.txt')
     const authtoken = fs.readFileSync(filePath, 'utf8')
+    if(!authtoken){
+        console.log('SignIn to Remove Data')
+        return
+    }
     try {
         await axios.delete(removepath,{
             data:{"authtoken":authtoken,"id": id}
@@ -18,7 +22,7 @@ const Delete = async (id) => {
                 setTimeout(fun, 1000);
             })
     } catch (error) {
-        console.log(error);
+        console.log(error.response.data.msg);
         const fun = () => {
             process.exit(0);
         }
